NullInputStream: ludicrous speed :) ThrottledInputStream/ThrottledOutputStream: -using nano instead of ms -> smaller deviation from given limit -added simple tester -fixed speedlimiter
git-svn-id: svn://svn.appwork.org/utils@1807 21714237-3853-44ef-a1f0-ef8f03a7d1fe

+++ src/org/appwork/utils/Files.java
@@ -61,8 +61,11 @@ public class Files {
     public static void deleteRecursiv(final File file) throws IOException {
         if (!file.exists()) { throw new FileNotFoundException(file.getAbsolutePath()); }
         if (file.isDirectory()) {
-            for (final File f : file.listFiles()) {
-                Files.deleteRecursiv(f);
+            final File[] files = file.listFiles();
+            if (files != null) {
+                for (final File f : files) {
+                    Files.deleteRecursiv(f);
+                }
             }
         }
         final boolean fd = file.delete();
